requirements.txt
python-dateutil
lxml
python-jsonrpc
pymongo
pika
requests
redis
sklearn
